Add errorchecking to getAddresses()
By querying the database for subnet and section individually we get
better control and errorchecking since we now know if a result of zero
addresses means we did something wrong or if it's just an empty subnet.
Method now return undef if something went wrong, or an array with no
elements if it's just an empty subnet.